Move QueryContext to separate include
