+++ src/include/duckdb/storage/block_manager.hpp
@@ -12,6 +12,7 @@
 #include "duckdb/common/mutex.hpp"
 #include "duckdb/common/optional_idx.hpp"
 #include "duckdb/common/unordered_map.hpp"
+#include "duckdb/main/query_context.hpp"
 #include "duckdb/storage/block.hpp"
 #include "duckdb/storage/storage_info.hpp"
 
@@ -20,7 +21,6 @@ namespace duckdb {
 class BlockHandle;
 class BufferHandle;
 class BufferManager;
-class ClientContext;
 class DatabaseInstance;
 class MetadataManager;
